requirements.txt
fastapi
uvicorn
google-cloud-storage
opencv-python-headless
google-genai
